% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{cal_count}
\alias{cal_count}
\title{Count the number of business days}
\usage{
cal_count(starts, stops, cal = calendar())
}
\arguments{
\item{starts, stops}{\code{[Date]}

Vectors of \code{Date}s determining the boundaries to count business days
between. Recycling is performed using standard tidyverse recycling rules.}

\item{cal}{\code{[calendar]}

A calendar.}
}
\description{
\code{cal_count()} counts the number of business days between \code{starts} and
\code{stops}. The count is inclusive for \code{starts}, but not inclusive for \code{stops}.
In mathematical notation, it counts the business days in the range of
\code{[starts, stops)}.
}
\examples{

# - 2018-12-31 is a business day (starts are inclusive)
# - 2019-01-01 is a holiday
# - 2019-01-02 is a business day
# - 2019-01-02 is not included (stops are not inclusive)
cal_count("2018-12-31", "2019-01-03")

# Technically it is allowed for `starts` to be after `stops`
cal_count("2019-01-03", "2018-12-31")

}
