% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar.R
\docType{data}
\name{conventions}
\alias{conventions}
\title{Business day conventions}
\format{An object of class \code{vctrs_list_of} (inherits from \code{vctrs_vctr}) of length 7.}
\usage{
conventions
}
\description{
The following business day conventions are used throughout the package. They
are generally applied when you land on a holiday or weekend, and have to
decide which direction to shift to move to the next date.
\itemize{
\item \code{"following"}

Choose the first business day after the given holiday.
\item \code{"modified_following"}

Choose the first business day after the given
holiday unless it belongs to a different month, in which case choose the
first business day before the holiday.
\item \code{"preceding"}

Choose the first business day before the given holiday.
\item \code{"modified_preceding"}

Choose the first business day before the given holiday unless it belongs
to a different month, in which case choose the first business day after
the holiday.
\item \code{"unadjusted"}

No adjustment is made.
\item \code{"half_month_modified_following"}

Choose the first business day after the given holiday unless that day
crosses the mid-month (15th) or the end of month, in which case choose
the first business day before the holiday.
\item \code{"nearest"}

Choose the nearest business day to the given holiday. If both the
preceding and following business days are equally far away, default to
following business day.
}
}
\keyword{datasets}
