% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/holidays.R
\name{holidays_add}
\alias{holidays_add}
\alias{holidays_remove}
\title{Add and remove holidays}
\usage{
holidays_add(calendar, holidays)

holidays_remove(calendar, holidays)
}
\description{
\itemize{
\item Use \code{holidays_add()} to add new holidays to a calendar. Existing holidays
are ignored. If a holiday you are trying to add has previously been
removed, then it will be re-added.
\item Use \code{holiday_remove()} to remove holidays from a calendar. These can be
manually added holidays from using \code{holidays_add()}, or pre-existing
holidays in that calendar (such as July 4th for a certain year in the
\code{"united_states"} calendar). Dates that are not holidays are ignored.
}
}
