% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/holidays.R
\name{holidays_all}
\alias{holidays_all}
\alias{holidays_all_calendars}
\alias{holidays_between}
\title{Find holidays}
\usage{
holidays_all(cal, weekends = FALSE)

holidays_all_calendars(weekends = FALSE)

holidays_between(start, stop, weekends = FALSE, cal = calendar())
}
\arguments{
\item{cal}{\code{[calendar]}

A calendar.}

\item{weekends}{\code{[logical(1)]}

Should weekends be treated as holidays?}

\item{start, stop}{\code{[Date(1)]}

Boundary values to look for holidays between. Both are inclusive
boundaries. It is required that \code{start} is \emph{strictly} less than \code{stop}.}
}
\value{
\itemize{
\item \code{holidays_all()} returns a vector of \code{Date}s.
\item \code{holidays_all_calendars()} returns a data frame with a \code{character}
\code{"calendar"} column, and a \code{list_of<date>} \code{"holidays"} column. If
the tibble package is installed, a tibble will be returned.
\item \code{holidays_between()} returns a vector of \code{Date}s.
}
}
\description{
These functions help with locating \emph{all} holidays for a specific calendar.
They locate both pre-existing and manually added holidays.
\itemize{
\item \code{holidays_all()} lists every holiday in \code{cal} from the minimum date
(1901-01-01) to the maximum date (2199-12-30).
\item \code{holidays_all_calendars()} maps \code{holidays_all()} over all calendars
available in \link{calendars}, and returns a data frame.
\item \code{holidays_between()} locates holidays betwen two dates.
}
}
\examples{
us_holidays <- holidays_all(calendar())

length(us_holidays)
us_holidays[1:5]

holidays_all_calendars()

# Locate holidays between two dates
cal <- calendar()

holidays_between("2019-01-01", "2019-03-01", cal = cal)

# Manually added holidays are respected
cal <- holidays_add(cal, "2019-01-02")
holidays_between("2019-01-01", "2019-03-01", cal = cal)

# So are manually removed holidays
cal <- holidays_remove(cal, "2019-01-01")
holidays_between("2019-01-01", "2019-03-01", cal = cal)

}
