% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/almanac-package.R
\docType{package}
\name{almanac-package}
\alias{almanac}
\alias{almanac-package}
\title{almanac: Tools for Business Calendars}
\description{
Provides tools for creating business calendars with 
    custom holidays. Calendars can then be used to shift dates or generate 
    sequences relative to that calendar.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/DavisVaughan/almanac}
  \item Report bugs at \url{https://github.com/DavisVaughan/almanac/issues}
}

}
\author{
\strong{Maintainer}: Davis Vaughan \email{davis@rstudio.com}

}
\keyword{internal}
