% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dates.R
\name{cal_adjust}
\alias{cal_adjust}
\title{Adjust a date}
\usage{
cal_adjust(x, convention = conventions$following, cal = calendar())
}
\description{
\itemize{
\item \code{cal_adjust()} adjusts a sequence of dates according to a business
calendar. If a date is currently a holiday or weekend, it is adjusted
using the corresponding \code{convention}.
}
}
