% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendar.R
\docType{data}
\name{calendar}
\alias{calendar}
\alias{default_calendar}
\alias{set_default_calendar}
\alias{calendars}
\title{Construct a calendar}
\format{An object of class \code{list} of length 5.}
\usage{
calendar(name = calendars$united_states)

default_calendar()

set_default_calendar(x)

calendars
}
\description{
These are various functions revolving around creating calendar objects, and
setting the default calendar.

- `calendar()` constructs a new 'calendar' object.

- `default_calendar()` is the current default calendar. The default is just
  to use `calendar()`.

- `set_default_calendar()` is used to alter the default calendar. This is
  set globally for all future calls to `default_calendar()` in the current
  R session.

- `calendars` lists the set of possible calendars that can be used.
}
\keyword{datasets}
